Declare missing public APIs

+++ belle-sip/include/belle-sip/utils.h
@@ -202,11 +202,9 @@ BELLESIP_EXPORT int belle_sip_getaddrinfo(const char *node, const char *service,
 BELLESIP_EXPORT void belle_sip_freeaddrinfo(struct addrinfo *res);
 
 
-
 typedef void (*belle_sip_background_task_end_callback_t)(void *);
-
-BELLESIP_EXPORT unsigned long belle_sip_begin_background_task(const char *name, belle_sip_background_task_end_callback_t cb, void *data);
-BELLESIP_EXPORT void belle_sip_end_background_task(unsigned long id);
+BELLESIP_EXPORT unsigned int belle_sip_begin_background_task(const char *name, belle_sip_background_task_end_callback_t cb, void *data);
+BELLESIP_EXPORT void belle_sip_end_background_task(unsigned int id);
 
 BELLE_SIP_END_DECLS
